requirements.txt
aiogram==2.18
aiohttp==3.8.1
aioredis==2.0.1
aiosignal==1.2.0
async-timeout==4.0.2
asyncpg==0.25.0
attrs==21.4.0
Babel==2.9.1
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.10
environs==9.4.0
frozenlist==1.3.0
gino==1.0.1
google-api-core==2.4.0
google-api-python-client==2.36.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
gspread==4.0.1
gspread-asyncio==1.5.3
httplib2==0.20.2
idna==3.3
loguru==0.6.0
marshmallow==3.14.1
multidict==6.0.2
oauth2client==4.1.3
oauthlib==3.2.0
peewee==3.14.8
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
SQLAlchemy==1.3.24
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
yarl==1.7.2
